JunoTerra_Internship/Task-1 satim/requirements.txt
animate==1.1.4.7
certifi==2020.6.20
chardet==3.0.4
Click==7.0
ffmpeg==1.4
Flask==1.1.1
Flask-Cors==3.0.8
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
MarkupSafe==1.1.1
mercantile==1.1.6
matplotlib == 3.2.2
numpy==1.17.4
opencv-python-headless==4.2.0.32
Pillow==8.0.1
requests==2.25.0
six==1.13.0
urllib3==1.26.2
Werkzeug==0.16.0
